+++ scripts/constants.mjs
@@ -1,6 +1,6 @@
 export const MODULE_ID = "dnd5e-crafting-core";
 export const MODULE_TITLE = "Crafting Core (DnD 5e)";
-export const MODULE_VERSION = "0.0.19";
+export const MODULE_VERSION = "0.0.19d";
 
 export const SETTINGS = Object.freeze({
   RECIPES: "recipes",